requirements.txt
crewai>=0.30.0
crewai-tools>=0.2.0
langchain-openai>=0.1.7
python-dotenv>=1.0.1
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
yfinance>=0.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
